backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-bootstrap4==23.2
django-extensions==3.2.3
soupsieve==2.5
sqlparse==0.4.4
djangorestframework==3.14.0
django-filter==23.3
Markdown==3.4.4
django-cors-headers==4.2.0
django-filter==23.3
lorem-text==2.1